requirements.txt
streamlit>=1.32.0
openai>=1.0.0
pinecone
python-dotenv>=1.0.0
pandas>=2.2.0
requests>=2.31.0
beautifulsoup4>=4.12.3